Code review: tweak hook implementation
